refactor: tweak styling & reorganize settings content
