+++ src/components/list-lemplate/list-template.module.scss
@@ -23,9 +23,6 @@
 }
 
 .table {
-    background-color: $clr-secondary;
-    color: $clr-dark;
-
     &Wrapper {
         box-sizing: border-box;
         width: 100%;
