--- src/helpers/theme-provider/index.ts
@@ -1,1 +1,0 @@
-export { default as ThemeProvider } from "./theme-provider";
