+++ src/index.scss
@@ -1,7 +1,7 @@
 body {
     font-family: Roboto, Helvetica, Arial, sans-serif;
-    color: black;
-    background-color: white;
+    color: #2e2e2e;
+    background-color: #ffffff;
     font-weight: 400;
     margin: 0;
     padding: 0;
@@ -13,7 +13,6 @@ body {
 
 a,
 a:hover {
-    color: #2e2e2e;
     text-decoration: none;
 }
