+++ src/modules/auth/components/password-input/password-input.module.scss
@@ -20,6 +20,8 @@
             height: 1rem;
         }
 
-        color: $clr-dark;
+        &_dark {
+            color: $clr-dark;
+        }
     }
 }
