+++ src/modules/settings/components/setting-select/index.ts
