+++ src/modules/settings/components/settings-content/index.ts
@@ -1,0 +1,3 @@
+import { lazy } from "react";
+
+export const SettingsContent = lazy(() => import("./settings-content"));
