+++ src/pages/settings/settings.tsx
@@ -1,22 +1,14 @@
 import { useTranslation } from "react-i18next";
 
 import { useGenericBreadcrumbs } from "@/hooks";
-import { ThemeSelect } from "@/components/theme-select";
-import { LangSelect } from "@/components/lang-select";
-
-import styles from "./settings.module.scss";
+import { SettingsContent } from "@/modules/settings/components/settings-content";
 
 const SettingsPage = () => {
     const { t } = useTranslation();
 
     useGenericBreadcrumbs(t("Settings"));
 
-    return (
-        <div className={styles.wrapper}>
-            <ThemeSelect />
-            <LangSelect />
-        </div>
-    );
+    return <SettingsContent />;
 };
 
 export default SettingsPage;
