+++ src/services/theme-service/themes/index.ts
@@ -1,0 +1,2 @@
+export { darkTheme } from "./dark";
+export { lightTheme } from "./light";
